fix: incorrect init value of viewType
