+++ .changeset/four-deers-bake.md
@@ -1,0 +1,7 @@
+---
+'@sajari/react-hooks': patch
+'@sajari/react-search-ui': patch
+'@sajari/react-sdk-utils': patch
+---
+
+Fix incorrect intial value of viewType
