+++ packages/hooks/src/utils/queryParams.ts
@@ -14,17 +14,6 @@ export function paramToRange(value: string) {
   return value.split(':').map(Number);
 }
 
-export function getSearchParams() {
-  if (isSSR()) {
-    return {};
-  }
-  const search: Record<string, string> = {};
-  new URLSearchParams(window.location.search).forEach((value, key) => {
-    search[key] = value;
-  });
-  return search;
-}
-
 export const initFiltersFromURLState = ({
   filters,
   params,
