+++ packages/search-ui/src/utils/queryParams.ts
@@ -1,0 +1,5 @@
+import { ResultViewType } from '../ContextProvider/types';
+
+export function isViewType(value: string): value is ResultViewType {
+  return ['grid', 'list'].includes(value);
+}
