+++ packages/utils/src/search-params.ts
@@ -1,0 +1,12 @@
+import { isSSR } from './ssr';
+
+export function getSearchParams() {
+  if (isSSR()) {
+    return {};
+  }
+  const search: Record<string, string> = {};
+  new URLSearchParams(window.location.search).forEach((value, key) => {
+    search[key] = value;
+  });
+  return search;
+}
